requirements.txt
fuzzywuzzy
python-Levenshtein
pytesseract
Pillow
pyenchant
PyMuPDF
PyPDF2
py_trans
langdetect
fake_useragent
openai
edge_tts
vosk
EdgeGPT==0.8.0
telebot
pysocks
html2text
youtube_transcript_api
selenium
chardet
python-magic
